Add support for expanding PODS_TARGET_SRCROOT in CMake (#3267)
* Fix a few things that should have been splatted but weren't

* Define PODS_TARGET_SRCROOT in CMake for compatibility with CocoaPods

* Compile pods we depend upon with looser settings

* Rename the nanopb dependency to match what CMake actually has

* Make #include <nanopb/pb.h> work

Remove the old way of #include <pb.h> once we change the Firestore
nanopb generator to match.

* Fix warnings that fail the build

+++ GoogleDataTransportCCTSupport/GDTCCTLibrary/GDTCCTPrioritizer.m
@@ -21,7 +21,7 @@
 #import <GoogleDataTransport/GDTStoredEvent.h>
 #import <GoogleDataTransport/GDTTargets.h>
 
-const static NSUInteger kMillisPerDay = 8.64e+7;
+const static int64_t kMillisPerDay = 8.64e+7;
 
 @implementation GDTCCTPrioritizer
